CA-175383: fix upgrade when storage partition is not on the primary disk

+++ upgrade.py
@@ -191,11 +191,12 @@ class ThirdGenUpgrader(Upgrader):
 
                 tool.commit(log = True)
 
-                storage_part = partitionDevice(primary_disk, storage_partnum)
-                rc, self.vgs_output_wrong  = util.runCmd2(['vgs', '--noheadings', '-o', 'vg_name'], with_stdout = True)
-                self.vgs_output_wrong = self.vgs_output_wrong.strip()
-                util.runCmd2(['vgremove', '-f', self.vgs_output_wrong])
-                util.runCmd2(['vgcreate', self.vgs_output, storage_part])
+                if storage_partnum > 0:
+                    storage_part = partitionDevice(primary_disk, storage_partnum)
+                    rc, self.vgs_output_wrong  = util.runCmd2(['vgs', '--noheadings', '-o', 'vg_name'], with_stdout = True)
+                    self.vgs_output_wrong = self.vgs_output_wrong.strip()
+                    util.runCmd2(['vgremove', '-f', self.vgs_output_wrong])
+                    util.runCmd2(['vgcreate', self.vgs_output, storage_part])
 
             else:
 
@@ -238,10 +239,11 @@ class ThirdGenUpgrader(Upgrader):
             # Remove current Volume Group
             util.runCmd2(['vgremove', '-f', self.vgs_output])
             # Remove LVM Phisical Volume
-            storage_part = partitionDevice(target_disk, storage_partnum)
-            util.runCmd2(['pvremove', storage_part])
-            # Delete LVM partition
-            tool.deletePartition(storage_partnum)
+            if storage_partnum > 0:
+                storage_part = partitionDevice(target_disk, storage_partnum)
+                util.runCmd2(['pvremove', storage_part])
+                # Delete LVM partition
+                tool.deletePartition(storage_partnum)
             # Resize backup partition
             tool.resizePartition(number = backup_partnum, sizeBytes = constants.backup_size * 2**20)
             # Write partition table
